--- api-boot-project/api-boot-plugins/api-boot-plugin-alibaba-oss/pom.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>api-boot-plugins</artifactId>
-        <groupId>org.minbox.framework</groupId>
-        <version>${revision}</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-    <description>ApiBoot Plugin 阿里云 对象存储oss</description>
-    <artifactId>api-boot-plugin-alibaba-oss</artifactId>
-    <properties>
-        <main.basedir>${basedir}/../../..</main.basedir>
-    </properties>
-    <dependencies>
-        <dependency>
-            <groupId>org.minbox.framework</groupId>
-            <artifactId>api-boot-plugin</artifactId>
-        </dependency>
-        <!--alibaba oss-->
-        <dependency>
-            <groupId>com.aliyun.oss</groupId>
-            <artifactId>aliyun-sdk-oss</artifactId>
-        </dependency>
-    </dependencies>
-
-</project>
