+++ api-boot-project/api-boot/pom.xml
@@ -15,6 +15,12 @@
     <name>ApiBoot</name>
     <artifactId>api-boot</artifactId>
     <dependencies>
+        <!--alibaba oss-->
+        <dependency>
+            <groupId>com.aliyun.oss</groupId>
+            <artifactId>aliyun-sdk-oss</artifactId>
+            <optional>true</optional>
+        </dependency>
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
