+++ api-boot-project/api-boot/src/main/java/org/minbox/framework/api/boot/oss/PartSize.java
@@ -15,7 +15,7 @@
  *
  */
 
-package org.minbox.framework.api.boot.plugin.oss;
+package org.minbox.framework.api.boot.oss;
 
 /**
  * File size unit
